Fix IPFS id mismatches. Be more flexible with URIs.
We cannot set "fs:" as image id because opennebula does not like a ":"
in the paths it passes to the TM scripts (":" is used as separator between
host and path). Therefore we strip fs:/ when adding by PATH. Adding by
source does not run the "cp" script and therefore ipfs ids need to be
provided directly without "fs:" part, or things break.